backend/requirements.txt
fastapi
mangum
boto3
pydantic
uvicorn
python-dotenv
